moves non-component files to src folder
